refactor(besu-validator): move code to connector plugin
Resolves #363
