+++ packages/cactus-plugin-ledger-connector-besu/src/main/typescript/openapi-spec.ts
@@ -2,6 +2,7 @@ import { OpenAPIV3 } from "openapi-types";
 import { DeployContractSolidityBytecodeEndpoint } from "./web-services/deploy-contract-solidity-bytecode-endpoint-constants";
 import { RunTransactionEndpoint } from "./web-services/run-transaction-endpoint-constants";
 import { InvokeContractEndpoint } from "./web-services/invoke-contract-endpoint-constants";
+import { BesuSignTransactionEndpointV1 } from "./web-services/sign-transaction-endpoint-constants";
 
 export const CACTUS_OPEN_API_JSON: OpenAPIV3.Document = {
   openapi: "3.0.3",
@@ -473,6 +474,46 @@ export const CACTUS_OPEN_API_JSON: OpenAPIV3.Document = {
           callOutput: {},
         },
       },
+      SignTransactionRequest: {
+        type: "object",
+        required: ["transactionHash", "keychainId", "keychainRef"],
+        properties: {
+          keychainId: {
+            type: "string",
+            minLength: 1,
+            maxLength: 100,
+            nullable: false,
+          },
+          keychainRef: {
+            type: "string",
+            minLength: 1,
+            maxLength: 100,
+            nullable: false,
+          },
+          transactionHash: {
+            description:
+              "The transaction hash of ledger will be used to fetch the contain.",
+            type: "string",
+            minLength: 0,
+            maxLength: 2048,
+            nullable: false,
+          },
+        },
+      },
+      SignTransactionResponse: {
+        type: "object",
+        required: ["signature"],
+        properties: {
+          signature: {
+            description:
+              "The signatures of ledger from the corresponding transaction hash.",
+            type: "string",
+            minLength: 0,
+            maxLength: 2048,
+            nullable: false,
+          },
+        },
+      },
     },
   },
   paths: {
@@ -559,6 +600,42 @@ export const CACTUS_OPEN_API_JSON: OpenAPIV3.Document = {
                 },
               },
             },
+          },
+        },
+      },
+    },
+    [BesuSignTransactionEndpointV1.HTTP_PATH]: {
+      [BesuSignTransactionEndpointV1.HTTP_VERB_LOWER_CASE]: {
+        operationId: BesuSignTransactionEndpointV1.OPENAPI_OPERATION_ID,
+        summary:
+          "Obtain signatures of ledger from the corresponding transaction hash.",
+        description:
+          "Obtain signatures of ledger from the corresponding transaction hash.",
+        parameters: [],
+        requestBody: {
+          required: true,
+          content: {
+            "application/json": {
+              schema: {
+                $ref: "#/components/schemas/SignTransactionRequest",
+              },
+            },
+          },
+        },
+        responses: {
+          "200": {
+            description: "OK",
+            content: {
+              "application/json": {
+                schema: {
+                  $ref: "#/components/schemas/SignTransactionResponse",
+                },
+              },
+            },
+          },
+          "404": {
+            description:
+              "Not able to find the corresponding tranaction from the transaction hash",
           },
         },
       },
