+++ packages/cactus-plugin-ledger-connector-besu/src/main/typescript/web-services/sign-transaction-endpoint-v1.ts
@@ -17,6 +17,7 @@ import {
   IJsObjectSignerOptions,
   KeyConverter,
   KeyFormat,
+  Checks,
 } from "@hyperledger/cactus-common";
 
 import {
@@ -25,47 +26,21 @@ import {
 } from "../generated/openapi/typescript-axios/api";
 
 import { BesuSignTransactionEndpointV1 as Constants } from "./sign-transaction-endpoint-constants";
+import { PluginLedgerConnectorBesu } from "../plugin-ledger-connector-besu";
 
 export interface IBesuSignTransactionEndpointOptions {
-  rpcApiHttpHost: string;
-  keyPairPem: string;
+  connector: PluginLedgerConnectorBesu;
   logLevel?: LogLevelDesc;
 }
 
 export class BesuSignTransactionEndpointV1 implements IWebServiceEndpoint {
-  private readonly web3: Web3;
-  private readonly jsObjectSigner: JsObjectSigner;
   private readonly log: Logger;
 
   constructor(public readonly options: IBesuSignTransactionEndpointOptions) {
     const fnTag = "BesuSignTransactionEndpointV1#constructor()";
-    if (!options) {
-      throw new Error(`${fnTag} options falsy.`);
-    }
-    if (!options.rpcApiHttpHost) {
-      throw new Error(`${fnTag} options.rpcApiHttpHost falsy.`);
-    }
-    if (!options.keyPairPem) {
-      throw new Error(`${fnTag} options.keyPairPem falsy.`);
-    }
 
-    const keyConverter = new KeyConverter();
-    const convertRawPrivate = keyConverter.privateKeyAs(
-      options.keyPairPem,
-      KeyFormat.PEM,
-      KeyFormat.Raw
-    );
-    const jsObjectSignerOptions: IJsObjectSignerOptions = {
-      privateKey: convertRawPrivate,
-      logLevel: options.logLevel || "INFO",
-    };
-
-    this.jsObjectSigner = new JsObjectSigner(jsObjectSignerOptions);
-
-    const web3Provider = new Web3.providers.HttpProvider(
-      this.options.rpcApiHttpHost
-    );
-    this.web3 = new Web3(web3Provider);
+    Checks.truthy(options, `${fnTag} options`);
+    Checks.truthy(options.connector, `${fnTag} options.connector`);
 
     const label = "besu-sign-transaction-endpoint";
     const level = options.logLevel || "INFO";
@@ -96,9 +71,7 @@ export class BesuSignTransactionEndpointV1 implements IWebServiceEndpoint {
     try {
       const request: SignTransactionRequest = req.body as SignTransactionRequest;
 
-      const trxResponse: Optional<SignTransactionResponse> = await this.processRequest(
-        request
-      );
+      const trxResponse = await this.options.connector.signTransaction(request);
 
       if (trxResponse.isPresent()) {
         res.status(200);
@@ -115,23 +88,5 @@ export class BesuSignTransactionEndpointV1 implements IWebServiceEndpoint {
       res.statusMessage = ex.message;
       res.json({ error: ex.stack });
     }
-  }
-
-  private async processRequest(
-    request: SignTransactionRequest
-  ): Promise<Optional<SignTransactionResponse>> {
-    const transaction = await this.web3.eth.getTransaction(
-      request.transactionHash
-    );
-
-    if (transaction !== undefined && transaction !== null) {
-      const singData = this.jsObjectSigner.sign(transaction.input);
-      const signDataHex = Buffer.from(singData).toString("hex");
-
-      const resBody: SignTransactionResponse = { signature: signDataHex };
-      return Optional.ofNullable(resBody);
-    }
-
-    return Optional.empty();
   }
 }
