--- packages/cactus-plugin-validator-besu/.gitignore
@@ -1,1 +1,0 @@
-cactus-openapi-spec-plugin-validator-besu.json
