--- packages/cactus-plugin-validator-besu/cactus-openapi-spec-plugin-validator-besu.json
@@ -1,96 +1,0 @@
-{
-    "openapi": "3.0.3",
-    "info": {
-        "title": "Hyperledger Cactus Plugin - Validator Besu Web Service",
-        "description": "Obtain signatures of ledger state from Cactus nodes through the API .",
-        "version": "0.0.1"
-    },
-    "servers": [
-        {
-            "url": "https://www.cactus.stream/{basePath}",
-            "description": "Public test instance",
-            "variables": {
-                "basePath": {
-                    "default": ""
-                }
-            }
-        },
-        {
-            "url": "http://localhost:4000/{basePath}",
-            "description": "Local test instance",
-            "variables": {
-                "basePath": {
-                    "default": ""
-                }
-            }
-        }
-    ],
-    "components": {
-        "schemas": {
-            "SignTransactionRequest": {
-                "type": "object",
-                "required": [
-                    "transactionHash"
-                ],
-                "properties": {
-                    "transactionHash": {
-                        "description": "The transaction hash of ledger will be used to fetch the contain.",
-                        "type": "string",
-                        "minLength": 0,
-                        "maxLength": 2048,
-                        "nullable": false
-                    }
-                }
-            },
-            "SignTransactionResponse": {
-                "type": "object",
-                "required": [
-                    "signature"
-                ],
-                "properties": {
-                    "signature": {
-                        "description": "The signatures of ledger from the corresponding transaction hash.",
-                        "type": "string",
-                        "minLength": 0,
-                        "maxLength": 2048,
-                        "nullable": false
-                    }
-                }
-            }
-        }
-    },
-    "paths": {
-        "/api/v1/plugins/@hyperledger/cactus-plugin-validator-besu/sign-transaction": {
-            "post": {
-                "summary": "Obtain signatures of ledger from the corresponding transaction hash.",
-                "description": "Obtain signatures of ledger from the corresponding transaction hash.",
-                "parameters": [],
-                "requestBody": {
-                    "required": true,
-                    "content": {
-                        "application/json": {
-                            "schema": {
-                                "$ref": "#/components/schemas/SignTransactionRequest"
-                            }
-                        }
-                    }
-                },
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "content": {
-                            "application/json": {
-                                "schema": {
-                                    "$ref": "#/components/schemas/SignTransactionResponse"
-                                }
-                            }
-                        }
-                    },
-                    "404": {
-                        "description": "Not able to find the corresponding tranaction from the transaction hash"
-                    }
-                }
-            }
-        }
-    }
-}
