--- packages/cactus-plugin-validator-besu/src/main/typescript/index.web.ts
@@ -1,1 +1,0 @@
-export * from "./generated/openapi/typescript-axios/index";
