--- packages/cactus-plugin-validator-besu/src/main/typescript/public-api.ts
@@ -1,12 +1,0 @@
-export {
-  BesuSignTransactionEndpointV1 as PostBesuSignTransactionEndpointV1,
-  IBesuSignTransactionEndpointOptions as IPostBesuSignTransactionEndpointOptions,
-} from "./besu/sign-transaction-endpoint-v1";
-
-export {
-  PluginValidatorBesu,
-  IPluginValidatorBesuOptions,
-  IWebAppOptions,
-} from "./plugin-validator-besu";
-
-export * from "./generated/openapi/typescript-axios/index";
