--- packages/cactus-plugin-validator-besu/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-      "outDir": "./dist/lib/", /* Redirect output structure to the directory. */
-      "declarationDir": "dist/types",
-    },
-    "include": [
-      "./src"
-    ]
-  }
