+++ packages/cactus-test-plugin-ledger-connector-besu/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "@hyperledger/cactus-test-plugin-validator-besu",
+  "name": "@hyperledger/cactus-test-plugin-ledger-connector-besu",
   "version": "0.2.0",
-  "description": "Integration tests for the Quorum ledger and the API server.",
+  "description": "Integration tests for the Besu ledger and the API server.",
   "main": "dist/lib/main/typescript/index.js",
-  "mainMinified": "dist/cactus-test-plugin-validator-besu.node.umd.min.js",
-  "browser": "dist/cactus-test-plugin-validator-besu.web.umd.js",
-  "browserMinified": "dist/cactus-test-plugin-cvalidator-besu.web.umd.min.js",
+  "mainMinified": "dist/cactus-test-plugin-ledger-connector-besu.node.umd.min.js",
+  "browser": "dist/cactus-test-plugin-ledger-connector-besu.web.umd.js",
+  "browserMinified": "dist/cactus-test-plugin-ledger-connector-besu.web.umd.min.js",
   "module": "dist/lib/main/typescript/index.js",
   "types": "dist/types/main/typescript/index.d.ts",
   "files": [
@@ -65,7 +65,8 @@
     "@hyperledger/cactus-cmd-api-server": "^0.2.0",
     "@hyperledger/cactus-common": "^0.2.0",
     "@hyperledger/cactus-core-api": "^0.2.0",
-    "@hyperledger/cactus-plugin-validator-besu": "^0.2.0",
+    "@hyperledger/cactus-plugin-keychain-memory": "^0.2.0",
+    "@hyperledger/cactus-plugin-ledger-connector-besu": "^0.2.0",
     "@hyperledger/cactus-test-tooling": "^0.2.0",
     "key-encoder": "2.0.3",
     "web3": "1.2.11",
