--- packages/cactus-test-plugin-validator-besu/src/main/typescript/index.ts
@@ -1,1 +1,0 @@
-export * from "./public-api";
