--- packages/cactus-test-plugin-validator-besu/src/test/typescript/integration/api-surface.ts
@@ -1,10 +1,0 @@
-// tslint:disable-next-line: no-var-requires
-const tap = require("tap");
-import * as publicApi from "../../../main/typescript/public-api";
-
-tap.pass("Test file can be executed");
-
-tap.test("Library can be loaded", (assert: any) => {
-  assert.plan(1);
-  assert.ok(publicApi);
-});
